# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "Smm Api Documentation",
        "contact": {
            "email": "takahiroasiro@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/admin/users": {
            "get": {
                "tags": [
                    "Admin/User"
                ],
                "summary": "Find user by name or email",
                "description": "Администратор: поиск пользователя по имени или email",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "The name parameter in query",
                        "required": true
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The email parameter in query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/admin/find": {
            "get": {
                "tags": [
                    "Admin/User"
                ],
                "summary": "Find user by name or email with balances",
                "description": "Администратор: поиск пользователя по имени или emailс их балансами",
                "operationId": "find",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "The name parameter in query",
                        "required": true
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The email parameter in query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
...
        "/api/user/transactions": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "User's transactions",
                "description": "Список транзакций пользователя",
                "operationId": "transactions",
                "parameters": [
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "The offset parameter in query",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "The limit parameter in query",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "The date_from parameter in query",
                        "required": false,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "The date_to parameter in query",
                        "required": false,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "transaction_group",
                        "in": "query",
                        "description": "The transaction_group parameter in query",
                        "required": false
                    },
                    {
                        "name": "cur",
                        "in": "query",
                        "description": "The cur parameter in query",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}